College scorecard & testing + Admin UI (#9)
* Base Code

* Remove node_modules and update git ignore

* Sequelize Base with connection to MySQL

* Add in the home page

* Replace Jade with Pug

* Be able to create user

* Completed back-end implementation

* Finished create account form

* Minor css changes

* Restructure directories

* Change view engine to pug

* Login & session (#2)

* Back-end login implementation with JWT

* Fix file path in auth.js

* Add login page & change style of create page

* Fix back-end bcrypt promise

* Add logout ability & navbar change

* modded gitignore to ignore config and c4me keys

* Added the RSA private/public keys

* Auto-add the admin user

* Modify validation of JWT

* Fix login cookies

* Minor fix for correcting function name

* College model (#3)

* Create college model and rank scraping

* Update college model and create major model

* Add scrape collegedata.com export and route

* Add scraping of majors and error catching

* Fix many to many relationship

* Change getadmin to checkadmin

* Add authentication to scrape routes

* Minor changes in values


* Add primary key for Major to not add dupes

* View & Edit Student Profile (#4)

* View Student Profile

* React bootrap items

* Edit Profile front-end load

* Edit Profile back-end & front-end

* Change the workings of applications + authentication fixes

* Comment on Application

* Add & update applications (front & back)

* Fixed loading and redirecting when trying to edit another user

* On create, go to login page

* Fix edit profile button

* Fix minor errors in profile page & updating

* Change const to let

* Remove loaded & filter existing applications

* toString all values in CollegeDropdown

* Added padding to bottom of app & null fix in CollegeDropdown

* Null issue in applications

* Complete import college scorecard

* Complete front end for scraping/importing

* Add delete students to admin front end

* Minor changes to fix error sending and validation

* Add import student profile to admin page

* Complete admin controls page

* Add testing to import college scorecard

* Minor changes to UI

* Add admin success alert

+++ back-end/models/college.js
@@ -21,7 +21,7 @@ module.exports = (sequelize, DataTypes) => {
             defaultValue: null
         },
         AdmissionRate: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(5, 2),
             validate: {
                 min: 0,
                 max: 100
@@ -42,7 +42,7 @@ module.exports = (sequelize, DataTypes) => {
             defaultValue: null
         },
         GPA: {
-            type: DataTypes.DECIMAL(10, 2),
+            type: DataTypes.DECIMAL(3, 2),
             validate: {
                 min: 0,
                 max: 4
@@ -78,7 +78,7 @@ module.exports = (sequelize, DataTypes) => {
             defaultValue: null
         },
         CompletionRate: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(5, 2),
             validate: {
                 min: 0,
                 max: 100
